// src/app/models/ticket-z.result.ts
export class TicketZResult {
    train: string;
    fromStationName: string;
    fromStation: string;
    toStationName: string;
    toStation: string;
    fromStationTime: string;
    toStationTime: string;
    take: string;
    businessClass: string;
    firstClass: string;
    secondClass: string;
    advancedSoftSleeper: string;
    softSleeper: string;
    moveSleeper: string;
    hardSleeper: string;
    softClass: string;
    hardClass: string;
    noSeat: string;
    others: string;
    remarks: string;

    trainNo: string;
    fromStationNo: string;
    toStationNo: string;
    seatTypes: string;
    trainDate: string;

}